Use result_t type for everything that can return error code Added exceptions to Java wrapper

+++ cdoc/IStreamSource.java
@@ -17,7 +17,7 @@ public class IStreamSource extends DataSource {
     }
 
     @Override
-    public int seek(long pos) {
+    public long seek(long pos) {
         System.err.format("IStreamSource: seek(%d)\n", pos);
         try {
             ifs.reset();
